+++ src/main/java/frc/robot/commands/common/arm/ArmCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm;
+package frc.robot.commands.common.arm;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj2.command.Command;
