+++ src/main/java/frc/robot/commands/common/drive/DriveCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.drive;
+package frc.robot.commands.common.drive;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
