+++ src/main/java/frc/robot/commands/common/shooter/SetShooterVelocity.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.shooter;
+package frc.robot.commands.common.shooter;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
