+++ src/main/java/frc/robot/commands/game/crescendo2024/auto/AutoIndexPiece.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.auto;
+package frc.robot.commands.game.crescendo2024.auto;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Subsystem;
