+++ src/main/java/frc/robot/commands/game/crescendo2024/auto/AutoNamedCommands.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.auto;
+package frc.robot.commands.game.crescendo2024.auto;
 
 import com.pathplanner.lib.auto.NamedCommands;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
