+++ src/main/java/frc/robot/commands/game/crescendo2024/auto/AutoPrepareForIntake.java
@@ -1,10 +1,10 @@
-package frc.robot.commands.auto;
+package frc.robot.commands.game.crescendo2024.auto;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.PrintCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
-import frc.robot.commands.arm.ArmToPosition;
+import frc.robot.commands.common.arm.ArmToPosition;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.arm.ArmSubsystem;
 import frc.robot.subsystems.intake.Intake;
