+++ src/main/java/frc/robot/commands/game/crescendo2024/auto/AutoPrepareForIntakeV2.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.auto;
+package frc.robot.commands.game.crescendo2024.auto;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj2.command.PrintCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
 import frc.robot.Constants;
-import frc.robot.commands.arm.ArmToPosition;
+import frc.robot.commands.common.arm.ArmToPosition;
 import frc.robot.config.RobotConfig;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.arm.ArmSubsystem;
