+++ src/main/java/frc/robot/commands/game/crescendo2024/debug/PrepareForScore.java
@@ -1,8 +1,8 @@
-package frc.robot.commands.debug;
+package frc.robot.commands.game.crescendo2024.debug;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.arm.ArmToPosition;
-import frc.robot.commands.shooter.SetShooterVelocity;
+import frc.robot.commands.common.arm.ArmToPosition;
+import frc.robot.commands.common.shooter.SetShooterVelocity;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.shooter.Shooter;
 import java.util.function.DoubleSupplier;
