+++ src/main/java/frc/robot/config/RobotConfigPhoenix.java
@@ -2,7 +2,7 @@ package frc.robot.config;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import edu.wpi.first.math.util.Units;
-import frc.robot.commands.auto.AutoNamedCommands;
+import frc.robot.commands.game.crescendo2024.auto.AutoNamedCommands;
 import frc.robot.subsystems.arm.ArmSubsystem;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.DriveBase;
