+++ src/main/java/frc/robot/controls/DebugControls.java
@@ -12,9 +12,9 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardLayout;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.commands.arm.ArmToPosition;
-import frc.robot.commands.debug.TestShooterAngle;
-import frc.robot.commands.shooter.SetShooterVelocity;
+import frc.robot.commands.common.arm.ArmToPosition;
+import frc.robot.commands.common.shooter.SetShooterVelocity;
+import frc.robot.commands.game.crescendo2024.debug.TestShooterAngle;
 import frc.robot.config.RobotConfig;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.arm.ArmSubsystem;
