+++ src/main/java/frc/robot/controls/DriverControls.java
@@ -26,11 +26,11 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.commands.arm.ArmCommand;
-import frc.robot.commands.assist.EjectPiece;
-import frc.robot.commands.drive.DriveCommand;
-import frc.robot.commands.drive.DriveToYaw;
-import frc.robot.commands.shooter.SetShooterVelocity;
+import frc.robot.commands.common.arm.ArmCommand;
+import frc.robot.commands.common.drive.DriveCommand;
+import frc.robot.commands.common.drive.DriveToYaw;
+import frc.robot.commands.common.shooter.SetShooterVelocity;
+import frc.robot.commands.game.crescendo2024.assist.EjectPiece;
 import frc.robot.config.RobotConfig;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.arm.ArmSubsystem;
